Added comments and one surprise :D
Added comments to explain certain lines that weren’t explained clearly
or were complicated.

Also added a fun segment that will check to see if the statement is
over 30 characters, and if it is, will say it fell asleep at around the
halfway mark.
For example, the statement:
“People say I talk too much so I try not to talk a lot but then I
always seem to ramble.”
will have the response:
“Sorry, I fell asleep around "People say I talk too much so I try not
to talk…””